Decouple SSZ scheme builder from SSZSerializer using interface SSZSchemeBuilder
